Issue19

+++ controllers/itempage.controller.js
@@ -1,4 +1,5 @@
 const ItemPageService = require('../services/itempage.service');
+const { itemPageModels, optionsModels, itemImgListsModels } = require('../validations/itempage.validation');
 
 class ItemPageController {
   itemPageService = new ItemPageService();
@@ -21,6 +22,9 @@ class ItemPageController {
 
     try {
       const oneItem = await this.itemPageService.findOneItem(itemId);
+      if (!oneItem) {
+        return res.status(400).json({ errorMessage: '상품이 존재하지 않습니다.' });
+      }
 
       res.status(200).json({ oneItem });
     } catch (err) {
@@ -31,8 +35,10 @@ class ItemPageController {
 
   // 상품 생성
   createItem = async (req, res) => {
-    let { brandName, itemName, rating, discount, price, colorData, sizeData, itemImgData } = req.body;
     const { userId } = res.locals.user;
+    const { brandName, itemName, rating, discount, price } = await itemPageModels.validateAsync(req.body);
+    const { colorData, sizeData } = await optionsModels.validateAsync(req.body);
+    const { itemImgData } = await itemImgListsModels.validateAsync(req.body);
 
     try {
       if (userId != '1') {
@@ -60,13 +66,20 @@ class ItemPageController {
 
   // 상품 수정
   updateItem = async (req, res) => {
-    const { brandName, itemName, rating, discount, price, colorData, sizeData, itemImgData } = req.body;
     const { itemId } = req.params;
     const { userId } = res.locals.user;
+    const { brandName, itemName, rating, discount, price } = await itemPageModels.validateAsync(req.body);
+    const { colorData, sizeData } = await optionsModels.validateAsync(req.body);
+    const { itemImgData } = await itemImgListsModels.validateAsync(req.body);
 
     try {
       if (userId != '1') {
-        res.status(400).json({ errorMessage: '상품 수정은 관리자만 가능합니다.' });
+        return res.status(400).json({ errorMessage: '상품 수정은 관리자만 가능합니다.' });
+      }
+
+      const updateTarget = await this.itemPageService.findOneItem(itemId);
+      if (!updateTarget) {
+        return res.status(400).json({ errorMessage: '수정할 상품이 존재하지 않습니다.' });
       }
 
       const updateItem = await this.itemPageService.updateItem(
@@ -97,6 +110,11 @@ class ItemPageController {
     try {
       if (userId != '1') {
         res.status(400).json({ errorMessage: '상품 삭제는 관리자만 가능합니다.' });
+      }
+
+      const deleteTarget = await this.itemPageService.findOneItem(itemId);
+      if (!deleteTarget) {
+        return res.status(400).json({ errorMessage: '삭제할 상품이 존재하지 않습니다.' });
       }
 
       const deleteItem = await this.itemPageService.deleteItem(userId, itemId);
